updated betalink function with important changes
Converts quantitative data to qualitative data with a warning. Also
makes it an important change to the calculation of beta_WN, beta_ST,
and b_contrib for the trivial case when no interactions are shared
between two networks. Specifically, all of these values are changed
from 0 to 1, which makes sense because these networks are totally
dissimilar due solely to species turnover.